staking: break out validator manager
